Create arithmetic_arranger
This function is to take five addition or subtraction problems or less and format it vertically and align them to the right. It, also, determines whether there is an operand other than '+' or '-' , only digits are in each problem, and the numbers are four digits or less. There is an optional argument to show the solutions to each problem, and its aligned automatically.